micropython/senml: Rename senml module.
MicroPython's senml module conflicts with https://pypi.org/project/senml/
when it's installed on PC (with pip) it overwrites/gets merged with pypi's
senml module, so it needs to be renamed.
