+++ micropython/senml/mpsenml/senml_base.py
